feat: re-use `SetupOptions.manifest` as much as possible
